Move the TLB utility to a new location
