+++ tests/atomspace/CMakeLists.txt
@@ -2,6 +2,7 @@
 LINK_LIBRARIES(
 	atombase
 	atomutils
+	atomspaceutils
 	atomspace
 	clearbox
 )
